% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plots.R
\name{gen_summary_data}
\alias{gen_summary_data}
\title{Generate the monthly summary data for a given site.}
\usage{
gen_summary_data(data_file, int_width = 0.8, model = 1, samples = 10,
  n_cores = 1)
}
\arguments{
\item{data_file}{An output file from \code{\parallel_logistic_sampler}.}

\item{int_width}{Width of the posterior probability interval.}

\item{model}{List index of the desired model.}

\item{samples}{Number of samples to take from the given model.}

\item{n_cores}{Number of cores to use in parallel.}
}
\value{
A list of the two summaries.
}
\description{
This is a wrapper for \code{\get_monthly_summaries}.
}
