% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plots.R
\name{gen_summary_data_mult}
\alias{gen_summary_data_mult}
\title{Generate the monthly summary data for multiple sites}
\usage{
gen_summary_data_mult(data_files, int_width = 0.8, model = 1,
  samples = 10, n_cores = 1)
}
\arguments{
\item{data_files}{A vector of datafiles to be read in.}

\item{int_width}{Width of the posterior probability interval.}

\item{model}{List index of the desired model.}

\item{samples}{Number of samples to take from the given model.}

\item{n_cores}{Number of cores to use in parallel.}
}
\value{
A list of the two summaries.
}
\description{
This function applies \code{\gen_summary_data} over many sites.
}
