% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{get_metadata}
\alias{get_metadata}
\title{Collect metadata for a given database.}
\usage{
get_metadata(database_file, site_nums = "all")
}
\arguments{
\item{database_file}{A file that contains the metadata and site data for the 
desired site, or is a destination for the desired database.}

\item{site_nums}{A vector of BOM site numbers.}
}
\value{
A dataframe of the site metadata.
}
\description{
Collect metadata for a given database.
}
