% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plots.R
\name{get_monthly_summaries}
\alias{get_monthly_summaries}
\title{Create the monthly summaries, for a given model.}
\usage{
get_monthly_summaries(sampler_results, y_sample, width, n_cores = 1)
}
\arguments{
\item{sampler_results}{a \code{ptsmlogistic} object.}

\item{y_sample}{A set of posterior predictive samples.}

\item{width}{The width of the posterior probability interval.}

\item{n_cores}{Number of cores desired to be used. Recommend setting this as high as you can.}
}
\description{
Create the monthly summaries, for a given model.
}
