% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{get_site_full}
\alias{get_site_full}
\title{Load site data.}
\usage{
get_site_full(database_file, site_num)
}
\arguments{
\item{database_file}{A file that contains the metadata and site data for the
desired site, or is a destination for the desired database.}

\item{site_num}{A BOM site number, e.g. 68016 for Cataract Dam.}
}
\value{
A dataframe of all daily rainfall measurements at \code{site_num}.
}
\description{
Load site data.
}
