% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_rainfall_data}
\alias{load_rainfall_data}
\title{Load the rainfall data for given site/s.}
\usage{
load_rainfall_data(db_file, sites)
}
\arguments{
\item{db_file}{A database file to be read from. If not present then it is created.}

\item{sites}{A vector of site numbers.}
}
\description{
Load the rainfall data for given site/s.
}
