% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-sampler.R
\name{parallel_marginal_sampler}
\alias{parallel_marginal_sampler}
\title{Run marginal MCMC sampler for the given site}
\usage{
parallel_marginal_sampler(site_data, iter, burn_in, min_components,
  max_components, output_dir = "./storm-output/", output_file = NA)
}
\arguments{
\item{site_data}{A dataframe containing the rainfall measurements. This
column must be called \code{rainfall}.}

\item{iter}{Number of sampling iterations.}

\item{burn_in}{Number of burn-in iterations.}

\item{min_components}{Minimum number of mixture components (>=2).}

\item{max_components}{Maximum number of mixture components.}

\item{output_dir}{The output directory for any output files.}

\item{output_file}{The output file}
}
\value{
A list of \code{ptsmlogistic} objects (one for each model ran).
}
\description{
Run marginal MCMC sampler for the given site
}
