% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plots.R
\name{plot_monthly_fits}
\alias{plot_monthly_fits}
\title{Plot the monthly values and the PPI's}
\usage{
plot_monthly_fits(monthly_data, start_date, end_date, scales = "fixed")
}
\arguments{
\item{monthly_data}{A dataframe of monthly}

\item{start_date}{The start date for the plot.}

\item{end_date}{The end date for the plot.}

\item{scales}{A string for what ggplot scale style to use, e.g.
\code{fixed}, \code{free_y}, etc.}
}
\description{
Plot the monthly values and the PPI's
}
