% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq-plots.R
\name{plot_qq_data}
\alias{plot_qq_data}
\title{Plot the qq data using ggplot}
\usage{
plot_qq_data(qq_data, legend_pos = "none")
}
\arguments{
\item{qq_data}{Output of \code{qq_data_period}.}

\item{legend_pos}{Legend to use (can be one of \code{top}, \code{bottom}, or \code{none}).}
}
\description{
Plot the qq data using ggplot
}
