% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly-plots.R
\name{print_proportion}
\alias{print_proportion}
\title{Print the actual proportion of points within the PPI.}
\usage{
print_proportion(summaries)
}
\arguments{
\item{summaries}{A dataframe that is the output from \code{\get_monthly_summaries}.}
}
\description{
Print the actual proportion of points within the PPI.
}
