% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq-plots.R
\name{qq_data_for_period}
\alias{qq_data_for_period}
\title{Compute the qq plot data from a \code{ptsmlogistic} objects.}
\usage{
qq_data_for_period(sampler_results, y_sample, months, period_name,
  probabilities)
}
\arguments{
\item{sampler_results}{A \code{ptsmlogistic} object.}

\item{y_sample}{Posterior predictive samples.}

\item{months}{Month numbers (as a vector of integers).}

\item{period_name}{The name of the season.}

\item{probabilities}{A vector of probabilities.}
}
\description{
Compute the qq plot data from a \code{ptsmlogistic} objects.
}
