% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq-plots.R
\name{qq_data_single}
\alias{qq_data_single}
\title{Get the qq data and compute it}
\usage{
qq_data_single(data_file, periods, probs, samples = 10, n_cores = 1)
}
\arguments{
\item{data_file}{The rds file to compute from.}

\item{periods}{The periods to compute from. Must be a list of lists.}

\item{probs}{Vector of probabilities.}

\item{samples}{Number of samples to take.}

\item{n_cores}{Number of cores to use in \code{mclapply} (if available).}
}
\description{
Get the qq data and compute it
}
