% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{rainfall_data_build}
\alias{rainfall_data_build}
\title{Build a dataframe with explanatory variables}
\usage{
rainfall_data_build(site_data, min_year = 0, min_month = 0)
}
\arguments{
\item{site_data}{A dataframe of daily rainfall observations for a single
site.}

\item{min_year}{The first year from which measurements should be taken.}

\item{min_month}{The first month (in \code{min_year}) from which
observations should be taken.}
}
\description{
Build a dataframe with explanatory variables
}
